Extract ResetStatement to reduce duplicate logic for MySQL and doris
